database/requirements.txt
python-dotenv==0.15.0
SQLAlchemy==1.4.31
sqlalchemy-utils==0.38.2
sqlalchemy-json==0.4.0
psycopg2==2.9.3